pyproject.toml
[tool.poetry]
name = "b3monitoringapi"
version = "0.1.0"
description = ""
authors = ["Maicon Renildo <maicon.renildo1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
peewee = "^3.15.1"
python-dotenv = "^0.20.0"
PyMySQL = "^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
